simple_lmk: Introduce Simple LMK (Low Memory Killer) for Android
This is a replacement for the Android Low Memory Killer driver that
simply targets a predefined number of pages to be freed every time a
memory allocation attempt is about to start killing processes, or
whenever it seems like kswapd is having trouble freeing enough memory
in a timely manner. Processes are killed according to the priorities
that Android gives them; in other words, the least important processes
are always killed first. Killing larger processes is preferred to
minimize the number of processes that need to be killed.

One way that memory reclaim is triggered is in the page-alloc slow path
after compaction has failed. When the Simple LMK memory reclaim process
is entered via the page allocator, a memory reclaim event is sent and
the caller is blocked until pages are freed and the memory allocation is
fulfilled. During this process, Simple LMK ensures that memory reclaim
cannot be invoked more than every CONFIG_ANDROID_SIMPLE_LMK_OOM_TIMEOUT
milliseconds. These OOM'd memory allocations are placed on a queue and
are prioritized over non-OOM'd memory allocations when new pages are
freed. This is to ensure that each OOM'd memory allocation gets
fulfilled as quickly as possible.

The other way that memory reclaim occurs is via kswapd. When kswapd runs
for longer than CONFIG_ANDROID_SIMPLE_LMK_KSWAPD_TIMEOUT milliseconds
without sleeping, then Simple LMK ensures that memory reclaim is invoked
at least once every CONFIG_ANDROID_SIMPLE_LMK_KSWAPD_TIMEOUT
milliseconds.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>